[Plat-731] use a struct as a key for token map
Collisions could occur if an exceptionally clever or (un)lucky user with
a particular name on a particular realm exist.

Example: user "pirtle" on realm "magic" stores his token in the map on
  the same key as user "irtle" on realm "mgaicp"

Though I think it'd only throw random auth errors, something tells me
someone could exploit this for token access to something they shouldn't
be able to see. Regardless! It's a bug and it's resolved in this commit.